backend/books_service/requirements.txt
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
coverage==7.3.2
Django==4.2.7
django-cors-headers==4.9.0
djangorestframework==3.14.0
idna==3.11
iniconfig==2.3.0
mysqlclient==2.1.1
packaging==25.0
pluggy==1.6.0
PyJWT==2.8.0
pytest==7.4.3
pytest-cov==4.1.0
pytest-django==4.5.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2025.2
requests==2.31.0
sqlparse==0.5.3
urllib3==2.5.0
pika==1.3.2
